Close #34 and #27 - convexity of tessellation and handling points outside [1,2]x[1,2]
Non-convexity of the tessellation means we are missing triangles.
A non-convex tessellation is obtained, if at least one of the corner points of the initial square lies inside the circumcircle of a missing triangle.
This problem is solved by scaling and shifting the points such that the union of the circumcircles of the triangles along the boundary of the tessellation lies inside (1,2)x(1,2).
With this modification it is now possible to start off with a point set outside [1,2]x[1,2].
Apply the function scaleShiftPoints to the point set before the tessellation and the function expand to the edge vertices after the tessellation.